LRAC-10551 Simplify
