--- modules/apps/analytics/analytics-batch-export-import-api/src/main/resources/com/liferay/analytics/batch/exportimport/client/packageinfo
@@ -1,1 +1,0 @@
-version 1.0.0
